requirements.txt
certifi==2026.1.4
charset-normalizer==3.4.4
dotenv==0.9.9
idna==3.11
psycopg2-binary==2.9.11
python-dotenv==1.2.2
requests==2.32.5
tzdata==2025.2
urllib3==2.6.3
